Moved code generation default class to external file
